OOFB-290 Ajusta tipo da variavel no script de estoque de consentimentos

+++ ferramentas-auxiliares/retirada-de-informaçoes-sobre-os-consentimentos/attachments/consent_function_receptor_stock.sql
@@ -1,7 +1,7 @@
 CREATE OR REPLACE FUNCTION consent_receptor_stock()
     RETURNS TABLE (
         org_name VARCHAR,
-        qtd_Estoque_Consentimentos_Ativos INT
+        qtd_Estoque_Consentimentos_Ativos BIGINT
 ) 
 LANGUAGE SQL
 AS $$
